requirements.txt
praw==6.5.1
requests==2.22.0
PyYAML==5.3
